🔄 Refactor : detach server component
